requirements.txt
agate==1.6.1
arrow==1.1.1
asn1crypto==1.4.0
asttokens==2.0.5
attrs==21.4.0
autopep8==1.6.0
azure-common==1.1.27
azure-core==1.21.1
azure-storage-blob==12.9.0
Babel==2.9.1
backcall==0.2.0
backoff==2.1.2
bazel==0.0.0.20200723
binaryornot==0.4.4
boto3==1.20.47
botocore==1.23.47
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
colored==1.4.3
cookiecutter==1.7.3
cryptography==3.4.8
dbt-bigquery==1.0.0
dbt-core==1.0.1
dbt-extractor==0.4.0
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
Gooey==1.0.8.1
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
google-cloud==0.34.0
google-cloud-bigquery==2.32.0
google-cloud-core==1.7.2
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.56.3
greenlet==2.0.1
grpcio==1.43.0
gspread==5.7.1
hologram==0.0.14
httplib2==0.20.4
idna==3.2
importlib-metadata==4.10.1
ipykernel==6.9.2
ipython==8.1.1
isodate==0.6.1
jedi==0.18.1
Jinja2==2.11.3
jinja2-time==0.2.0
jmespath==0.10.0
json-rpc==1.13.0
jsonschema==3.1.1
jupyter-client==7.1.2
jupyter-core==4.9.2
keyring==21.8.0
leather==0.3.4
Logbook==1.5.3
MarkupSafe==2.0.1
mashumaro==2.9
matplotlib-inline==0.1.3
minimal-snowplow-tracker==0.0.2
mpi4py==3.0.3
msgpack==1.0.3
msrest==0.6.21
nest-asyncio==1.5.4
networkx==2.6.3
numpy==1.22.4
oauthlib==3.2.0
openpyxl==3.0.10
oscrypto==1.2.1
packaging==20.9
pandas==1.2.5
parsedatetime==2.6
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
poyo==0.5.0
prompt-toolkit==3.0.28
proto-plus==1.19.9
protobuf==3.19.4
psutil==5.9.0
psycopg2-binary==2.9.3
pure-eval==0.2.2
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodomex==3.14.0
Pygments==2.11.2
pygtrie==2.4.2
PyJWT==2.3.0
pyOpenSSL==20.0.1
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-slugify==5.0.2
pytimeparse==1.1.8
pytz==2021.3
pywin32==303
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.1
six==1.16.0
snowflake-connector-python==2.5.1
SQLAlchemy==1.4.44
sqlparse==0.4.2
stack-data==0.2.0
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traitlets==5.1.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.2
wxPython==4.1.1
zipp==3.7.0
